requirements.txt
numpy==1.24.2
opencv-python==4.11.0.86
ultralytics==8.3.130
torch==2.1.0
scipy==1.15.3
pyttsx3==2.98
SpeechRecognition==3.14.2
vosk==0.3.45
pyaudio==0.2.13
ahrs==0.3.1
icm20948==1.0.0
scikit-image
filterpy
onnx
psutil==5.9.4